fix(stella-store): move the orphan-id reader out of a god file at its ceiling
`file size ratchet` was the only red check on this branch: `usage.rs` grew to
1832 lines against a recorded ceiling of 1767. That file is grandfathered and
closed to growth, and it was already sitting one line under its ceiling on
`main` (1766), so any addition to it fails by construction.

`make file-size-update` is the wrong remedy here. AGENTS.md reserves it for a
genuinely irreducible line, and 66 lines of new method plus its witness is not
that — it is the split the guard exists to ask for.

`orphan_project_ids` and its witness move to `usage/orphans.rs`, beside the
`usage/rekey.rs` sibling that already holds a table-spanning hub operation and
its own tests. Behaviour is unchanged: same SQL, same witness, same assertions.
`usage.rs` keeps the one line the split costs — the module declaration, placed
where rustfmt's ordering already wants it — landing at 1767, its ceiling, with
no baseline edit.

The module doc says at the site what the method's doc comment used to have to
compress: the two paths that produce an orphan, why the gone-checkout scan
cannot reach one, and that the org-row guard still applies.

The witness borrows `source_row` from the parent's fixtures (widening it costs
no line) but defines its own `scope`: widening the parent's reflows the
signature across four lines, which would breach the ceiling this commit is
clearing.

Verified on this branch:

  cargo test -p stella-store --lib usage::orphans   1 passed (the witness, in
                                                    its new home)
  cargo test -p stella-store --lib usage::          42 passed, 0 failed
  cargo clippy -p stella-store --all-targets -D warnings   clean
  cargo fmt --check -p stella-store                        clean
  RUSTDOCFLAGS=-D warnings cargo doc -p stella-store --document-private-items
                                                           clean
  ./scripts/check-file-size.sh                             exit 0

Targeted rather than workspace-wide, per SCR-001; CI runs the full gate.

Refs #6100
